custom hook for tab scroll positions
