+++ client/src/ER/ER.js
@@ -14,13 +14,17 @@ import Beds from './Beds';
 import Ringdowns from './Ringdowns';
 
 import notification from '../assets/notification.mp3';
+import { useTabPositions } from '../Components/SelectedTab';
 
 export default function ER() {
   const { hospitalUser } = useContext(Context);
   const socketUrl = `${window.location.origin.replace(/^http/, 'ws')}/wss/hospital?id=${hospitalUser?.hospital.id}`;
   const { lastMessage } = useWebSocket(socketUrl, { shouldReconnect: () => true });
+  const {selectedTab,handleSelectTab } = useTabPositions('ringdown', {
+    ringdown: 0,
+    hospitalInfo: 0,
+  });
 
-  const [selectedTab, setSelectedTab] = useState('ringdown');
   const [ringdowns, setRingdowns] = useState([]);
   const [unconfirmedRingdowns, setUnconfirmedRingdowns] = useState([]);
   const [statusUpdate, setStatusUpdate] = useState();
@@ -95,7 +99,7 @@ export default function ER() {
     <div className="grid-container minh-100vh">
       <div className="grid-row">
         <div className="tablet:grid-col-6 tablet:grid-offset-3">
-          <RoutedHeader selectedTab={selectedTab} onSelect={setSelectedTab} />
+          <RoutedHeader selectedTab={selectedTab} onSelect={handleSelectTab} />
           {showRingdown && (!showTabs || selectedTab === 'ringdown') && <Ringdowns ringdowns={ringdowns} onStatusChange={onStatusChange} />}
           {showInfo && (!showTabs || selectedTab === 'hospitalInfo') && (
             <Beds statusUpdate={statusUpdate} onStatusUpdate={onStatusUpdate} incomingRingdownsCount={incomingRingdownsCount} />
